pyproject.toml
[project]
name = "reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.6.0",
    "django-filter>=24.3",
    "python-dotenv>=1.0.1",
    "openai>=1.57.0",
]
